requirements.txt
dill
from-root
notebook
gdown
flask-cors
flask
matplotlib
numpy
pandas
geopandas
torchnet
ipykernel
wget
tqdm
GitPython
geopandas
-e .